+++ orders/views.py
@@ -1,5 +1,4 @@
 import json
-from enum                 import Enum
 
 from django.views         import View
 from django.http          import JsonResponse
@@ -7,26 +6,17 @@ from django.db            import transaction
 from django.db.models     import Sum, F
 from django.db.utils      import IntegrityError
 
-from orders.models        import Order, OrderItem, OrderStatus, OrderItemStatus, PaymentMethod
+from orders.models        import (
+        Order, 
+        OrderItem, 
+        OrderStatus, 
+        OrderItemStatus, 
+        PaymentMethod, 
+        OrderStatusEnum, 
+        OrderItemStatusEnum
+)
 from users.models         import Cart
 from users.utils          import login_decorator
-
-class OrderStatusEnum(Enum):
-    CONFIRMING      = 1
-    PAID            = 2
-    PENDING         = 3
-    DELIVERD        = 4
-    ORDER_CANCELLED = 5
-    REFUNDED        = 6
-
-class OrderItemStatusEnum(Enum):
-    PAID            = 1
-    PREPARING       = 2
-    SHIPPED         = 3
-    DELIVERD        = 4
-    ORDER_CANCELLED = 5
-    REFUNDED        = 6
-    RETURNED        = 7
 
 class OrderView(View):
     @login_decorator
@@ -40,14 +30,14 @@ class OrderView(View):
                 
                 order = Order.objects.create(
                     user            = user,
-                    status          = OrderStatus.objects.get(id = OrderStatusEnum.CONFIRMING.id),
+                    status          = OrderStatus.objects.get(id = OrderStatusEnum.CONFIRMING),
                     payment_method  = PaymentMethod.objects.get(name = payment_method)
                 )
                 
                 OrderItem.objects.bulk_create([OrderItem(
                         product  = cart.product,
                         order    = order,
-                        status   = OrderItemStatus.objects.get(id = OrderItemStatusEnum.PAID.id),
+                        status   = OrderItemStatus.objects.get(id = OrderItemStatusEnum.PAID),
                         quantity = cart.quantity,
                         price    = cart.product.price,
                     )for cart in carts])
@@ -91,10 +81,10 @@ class OrderView(View):
             with transaction.atomic():         
                 order    = Order.objects.prefetch_related("order_items").get(user_id = request.user.id, id = order_id)   
                 
-                order.status = OrderStatus.objects.get(id = OrderStatusEnum.ORDER_CANCELLED.id)
+                order.status = OrderStatus.objects.get(id = OrderStatusEnum.ORDER_CANCELLED)
                 order.save()
                 
-                order.order_items.all().update(status = OrderItemStatus.objects.get(id = OrderItemStatusEnum.ORDER_CANCELLED.id))
+                order.order_items.all().update(status = OrderItemStatus.objects.get(id = OrderItemStatusEnum.ORDER_CANCELLED))
                 
             return JsonResponse({'message': 'SUCCESS'}, status=200)
